back:wakaama, include block1.c, er-coap-13.c, packet.c, lwm2m_utils.c
 see RestoreLog:
1)  restore directly

+++ iot_link/network/lwm2m/lwm2m_imp/wakaama/wakaama-master/core/block1.c
@@ -160,7 +160,7 @@ uint8_t coap_block1_handler(lwm2m_block1_data_t ** pBlock1Data,
 
     if (blockMore)
     {
-        *outputLength = (size_t) - 1;
+        *outputLength = -1;
         return COAP_231_CONTINUE;
     }
     else
